refactor: 지도 페이지 하단 영역 분리 및 dynamic import
